Moved new MCTS logic into group14 folder and MyAgent.py file
