--- agents/Group4/cmd.txt
@@ -1,1 +1,0 @@
-agents.Group4.MyAgent_Ale MyAgent_Ale
